feat: implement user role assignment and authorization policies

+++ EventPlanner.Application/Workshops/Commands/DeleteWorkshop/DeleteWorkshopCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using EventPlanner.Domain.Constants;
 using EventPlanner.Domain.Entities;
 using EventPlanner.Domain.Exceptions;
+using EventPlanner.Domain.Interfaces;
 using EventPlanner.Domain.Repositories;
 using MediatR;
 using Microsoft.Extensions.Logging;
@@ -8,7 +10,8 @@ using Microsoft.Extensions.Logging;
 namespace EventPlanner.Application.Workshops.Commands.DeleteWorkshop;
 
 public class DeleteWorkshopCommandHandler(ILogger<DeleteWorkshopCommandHandler> logger,
-    IWorkshopsRepository workshopsRepository) : IRequestHandler<DeleteWorkshopCommand>
+    IWorkshopsRepository workshopsRepository,
+    IEventPlannerAuthorizationService eventPlannerAuthorizationService) : IRequestHandler<DeleteWorkshopCommand>
 {
     public async Task Handle(DeleteWorkshopCommand request, CancellationToken cancellationToken)
     {
@@ -17,6 +20,9 @@ public class DeleteWorkshopCommandHandler(ILogger<DeleteWorkshopCommandHandler> 
         var workshop = await workshopsRepository.GetByIdAsync(request.Id);
         if (workshop is null)
             throw new NotFoundException(nameof(Workshop), request.Id.ToString());
+
+        if (!eventPlannerAuthorizationService.Authorize(workshop, ResourceOperation.Delete))
+            throw new ForbidException();
 
         await workshopsRepository.Delete(workshop);
     }
